+++ src/renderer_d3d11.cpp
@@ -4023,7 +4023,7 @@ namespace bgfx { namespace d3d11
 			}
 		}
 
-		uint16_t shaderSize;
+		uint32_t shaderSize;
 		bx::read(&reader, shaderSize);
 
 		const void* code = reader.getDataPtr();
